requirements.txt
anyio==3.5.0
asgiref==3.4.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.71.0
fastapi-utils==0.2.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
motor==2.5.1
oauthlib==3.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pymongo==3.12.3
python-decouple==3.5
python-jose==3.3.0
python-multipart==0.0.5
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.29
starlette==0.17.1
tweepy==4.4.0
typing-extensions==4.0.1
urllib3==1.26.8
uvicorn==0.16.0
